Refactored code and modified delete integration

+++ lib/web/views.py
@@ -41,11 +41,11 @@ def userdetails(request):
     # Returns the integration id,user_invoke_name for an integration
     if request.user.is_authenticated:
         user_integrations = YellowUserToken.objects.filter(user=request.user.id)
-        print(user_integrations)
+        # print(user_integrations)
         for user_integration in user_integrations:
             try:
                 vout = VictorOpsUserToken.objects.get(user_integration=user_integration)
-                print(vout)
+                # print(vout)
                 user_integrations_list.append({"user_invoke_name": user_integration.\
                                               yellowant_integration_invoke_name,
                                                "id": user_integration.id, "app_authenticated": True,
@@ -75,21 +75,22 @@ def user_detail_update_delete_view(request, id=None):
     elif request.method == "DELETE":
         # deletes the integration
         access_token_dict = YellowUserToken.objects.get(id=id)
-        access_token = access_token_dict.yellowant_token
-        print(access_token)
-        user_integration_id = access_token_dict.yellowant_integration_id
-        print(user_integration_id)
-        url = "https://api.yellowant.com/api/user/integration/%s"%(user_integration_id)
-        yellowant_user = YellowAnt(access_token=access_token)
-        print(yellowant_user)
-        yellowant_user.delete_user_integration(id=user_integration_id)
-        response_json = YellowUserToken.objects.get(yellowant_token=access_token).delete()
-        print(response_json)
-        return HttpResponse("successResponse", status=204)
+        user_id = access_token_dict.user
+        if user_id == request.user.id:
+            access_token = access_token_dict.yellowant_token
+            user_integration_id = access_token_dict.yellowant_integration_id
+            url = "https://api.yellowant.com/api/user/integration/%s" % (user_integration_id)
+            yellowant_user = YellowAnt(access_token=access_token)
+            yellowant_user.delete_user_integration(id=user_integration_id)
+            response_json = YellowUserToken.objects.get(yellowant_token=access_token).delete()
+            # print(response_json)
+            return HttpResponse("successResponse", status=204)
+        else:
+            return HttpResponse("Not Authenticated", status=403)
     elif request.method == "PUT":
         # adds a new integration
         data = json.loads(request.body.decode("utf-8"))
-        print(data)
+        # print(data)
         user_id = data['user_id']
         api_id = data['victorops_api_id']
         api_key = data['victorops_api_key']
@@ -109,16 +110,14 @@ def user_detail_update_delete_view(request, id=None):
                 response.status_code = 200
 
         if response.status_code == 200:
-            print("Valid")
+            # print("Valid")
             api_new = VictorOpsUserToken.objects.get(user_integration_id=user_integration)
             api_new.victorops_user_id = user_id
             api_new.victorops_api_id = api_id
             api_new.victorops_api_key = api_key
             api_new.apikey_login_update_flag = True
             api_new.save()
-            print(api_new.victorops_api_id)
-            print(api_new.victorops_api_key)
             return HttpResponse("Submitted", status=200)
         else:
-            print("Invalid")
+            # print("Invalid")
             return HttpResponse("Invalid Credentials", status=401)
